Uploading 4th tutorial
In this tutorial we learn about connection resiliency and interception.  Connection resiliency is an app's ability to recognize transient error and continue attempting the action that generated them a certain number of times every so often.  Interception can catch SQL statement before they get to the db server.  Once intercepted the statements can be changed and/or logged.